+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fixed bug in `current_user_can_download=""` and `current_user_bought_product=""` attributes that triggering syntax errors in some rare scenarios.
 - Fixed bug that was causing an `E_NOTICE` level warning about `$limit`.
 - Fixed bug causing a PHP warning in some scenarios related to `s::setTransient()`.
+- Fixed bug that was causing `current_user_can_download=""` to fail when checking for a download sold by a product variation.
 
 ## v160720.23954